JigsawPiece class to hold all piece related logic. Config update.
This is a major restructure. JigsawPiece is now a class of its own with all the behavior implicit to a piece encapsulated in this class including the glow effect. The glow effect is slightly better in that it's now behind each piece rather than on top of it.

There's also a little update to the webpack config to allow importing without the `.ts` extensions.